Create SumByCriteria
This macro calculates the sum of values from the rngValues range where the corresponding cell in the rngCriteria range matches the value in cell B2. The calculated sum is then displayed in cell C2. The macro can be executed from the VBA editor within Excel or assigned to a button in the worksheet interface.